Fix homeWorld not updated when changing home
The homeWorld variable is checked when the player does /f home to teleport to his faction home. The thing is that it's always null until the server reboots. So if the faction is new and the home is new the player cannot teleport to it.
I don't really understand why homeWorld exists but maybe it's usefull so here is a fix.